Fix erroneous results when more search result pages are requested than are returned
