+++ src/main/kotlin/CharacterSearch.kt
@@ -13,8 +13,6 @@ import cloud.drakon.ktlodestone.character.search.scrapeCharacterSearch
 import cloud.drakon.ktlodestone.exception.LodestoneException
 import cloud.drakon.ktlodestone.world.DataCenter
 import cloud.drakon.ktlodestone.world.World
-import io.ktor.client.call.body
-import io.ktor.client.request.get
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.future.future
@@ -45,38 +43,17 @@ suspend fun searchLodestoneCharacter(
     grandCompanies: Set<GrandCompanyName?>? = null,
     languages: Set<Language>? = null,
     pages: Byte = 1,
-) = if (pages == 1.toByte()) {
-    searchLodestoneCharacterPaginated(
-        name,
-        world,
-        dataCenter,
-        classJob,
-        race,
-        clan,
-        grandCompanies,
-        languages
-    )
-} else if (pages > 1.toByte()) {
-    buildList {
-        for (page in 1..pages) {
-            add(
-                searchLodestoneCharacterPaginated(
-                    name,
-                    world,
-                    dataCenter,
-                    classJob,
-                    race,
-                    clan,
-                    grandCompanies,
-                    languages,
-                    page
-                )
-            )
-        }
-    }.flatten()
-} else {
-    throw InvalidParameterException("`pages` must be at least 1.")
-}
+) = if (pages >= 1) scrapeCharacterSearch(
+    name,
+    world,
+    dataCenter,
+    classJob,
+    race,
+    clan,
+    grandCompanies,
+    languages,
+    pages,
+) else throw InvalidParameterException("`pages` must be at least 1.")
 
 /**
  * Searches for a character on *The Lodestone*.
@@ -118,62 +95,4 @@ fun searchLodestoneCharacterAsync(
         languages,
         pages,
     )
-}
-
-private suspend fun searchLodestoneCharacterPaginated(
-    name: String?,
-    world: World?,
-    dataCenter: DataCenter?,
-    classJob: ClassJob?,
-    race: Race?,
-    clan: Clan?,
-    grandCompanies: Set<GrandCompanyName?>?,
-    languages: Set<Language>?,
-    page: Int? = null,
-) = ktorClient.get("character/") {
-    url {
-        // We've already encoded this parameter
-        if (name != null) encodedParameters.append("q", name.replace(" ", "+"))
-
-        // If both [world] and [dataCenter] or just [world] are provided, use [world] as it's more specific, otherwise use [dataCenter]
-        if ((world != null && dataCenter != null) || world != null) {
-            parameters.append("worldname", world.name)
-        } else if (dataCenter != null) {
-            parameters.append("worldname", dataCenter.name)
-        }
-
-        if (classJob != null) parameters.append("classjob", classJob.name)
-
-        // If both [clan] and [race] or just [clan] are provided, use [clan] as it's more specific, otherwise use [race]
-        if ((clan != null && race != null) || clan != null) {
-            parameters.append("race_tribe", clan.name)
-        } else if (race != null) {
-            parameters.append("race_tribe", race.name)
-        }
-
-        grandCompanies?.forEach {
-            when (it) {
-                GrandCompanyName.MAELSTROM -> parameters.append("gcid", "1")
-                GrandCompanyName.ORDER_OF_THE_TWIN_ADDER -> parameters.append("gcid", "2")
-                GrandCompanyName.IMMORTAL_FLAMES -> parameters.append("gcid", "3")
-                null -> parameters.append("gcid", "0")
-            }
-        }
-
-        languages?.forEach {
-            when (it) {
-                Language.JAPANESE -> parameters.append("blog_lang", "ja")
-                Language.ENGLISH -> parameters.append("blog_lang", "en")
-                Language.GERMAN -> parameters.append("blog_lang", "de")
-                Language.FRENCH -> parameters.append("blog_lang", "fr")
-            }
-        }
-
-        if (page != null) parameters.append("page", "$page")
-    }
-}.let {
-    when (it.status.value) {
-        200 -> scrapeCharacterSearch(it.body())
-        else -> throw LodestoneException()
-    }
 }
